+++ extension-impl/registry-consul/src/main/java/com/alipay/sofa/rpc/registry/consul/common/ConsulURL.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.rpc.registry.common;
+package com.alipay.sofa.rpc.registry.consul.common;
 
 import com.alipay.sofa.rpc.common.RpcConstants;
 import com.alipay.sofa.rpc.common.utils.CommonUtils;
