+++ extension-impl/registry-consul/src/main/java/com/alipay/sofa/rpc/registry/consul/internal/ConsulManager.java
@@ -14,11 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.rpc.registry.internal;
+package com.alipay.sofa.rpc.registry.consul.internal;
 
 import com.alipay.sofa.rpc.common.struct.NamedThreadFactory;
-import com.alipay.sofa.rpc.registry.common.ConsulConstants;
-import com.alipay.sofa.rpc.registry.model.*;
+import com.alipay.sofa.rpc.registry.consul.common.ConsulConstants;
+import com.alipay.sofa.rpc.registry.consul.model.*;
 import com.ecwid.consul.v1.ConsulClient;
 import com.ecwid.consul.v1.QueryParams;
 import com.ecwid.consul.v1.Response;
