+++ extension-impl/registry-consul/src/main/java/com/alipay/sofa/rpc/registry/consul/model/NotifyListener.java
@@ -14,9 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.rpc.registry.model;
+package com.alipay.sofa.rpc.registry.consul.model;
 
-import com.alipay.sofa.rpc.registry.common.ConsulURL;
+import com.alipay.sofa.rpc.registry.consul.common.ConsulURL;
 
 import java.util.List;
