+++ extension-impl/registry-consul/src/main/resources/META-INF/services/sofa-rpc/com.alipay.sofa.rpc.registry.Registry
@@ -1,1 +1,1 @@
-consul=com.alipay.sofa.rpc.registry.ConsulRegistry
+consul=com.alipay.sofa.rpc.registry.consul.ConsulRegistry
